+++ Controllers/QuestionController.cs
@@ -1,7 +1,9 @@
 ﻿using IyiOlusAdminPanel.Models.Questions;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Newtonsoft.Json;
 using System.Net.Http;
+using System.Net.Http.Headers;
 using System.Security.Cryptography.X509Certificates;
 using System.Text.Json;
 using System.Text.Json.Serialization;
@@ -23,10 +25,21 @@ namespace IyiOlusAdminPanel.Controllers
         [Route("Index")]
         public async Task<IActionResult> Index(int pageIndex = 0, int PageSize = 20)
         {
+            var token = Request.Cookies["token"];
+            if (string.IsNullOrEmpty(token))
+            {
+                return RedirectToAction("Login");
+            }
+
             var client = _httpClientFactory.CreateClient();
+            client.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue("Bearer", token);
+
             var responseString = await client.GetStringAsync(Constants.ApiRoot + $"Questions?PageIndex={pageIndex}&PageSize={PageSize}");
+
+
             var apiResponse = JsonConvert.DeserializeObject<QuestionResponse>(responseString);
-            var questions = apiResponse.Items;  
+            var questions = apiResponse.Items;
+
             // ViewData ile pagination bilgilerini gönder
             ViewData["CurrentPage"] = apiResponse?.pageNumber ?? 1;
             ViewData["TotalPages"] = apiResponse?.totalPages ?? 1;
@@ -38,6 +51,7 @@ namespace IyiOlusAdminPanel.Controllers
         [Route("Create")]
         public IActionResult Create()
         {
+
             return View();
         }
 
@@ -45,7 +59,14 @@ namespace IyiOlusAdminPanel.Controllers
         [Route("Create")]
         public async Task<IActionResult> Create(CreateQuestionDto createQuestionDto)
         {
+            var token = Request.Cookies["token"];
+            if (string.IsNullOrEmpty(token))
+            {
+                return RedirectToAction("Login");
+            }
+
             var client = _httpClientFactory.CreateClient();
+            client.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue("Bearer", token);
             var payload = new
             {
                 request = createQuestionDto
@@ -60,7 +81,15 @@ namespace IyiOlusAdminPanel.Controllers
         [Route("Update/{id}")]
         public async Task<IActionResult> Update(Guid id)
         {
+            var token = Request.Cookies["token"];
+            if (string.IsNullOrEmpty(token))
+            {
+                return RedirectToAction("Login");
+            }
+
             var client = _httpClientFactory.CreateClient();
+            client.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue("Bearer", token);
+
             var response = await client.GetAsync(Constants.ApiRoot + $"Questions/{id}");
 
             var options = new JsonSerializerOptions
@@ -78,7 +107,15 @@ namespace IyiOlusAdminPanel.Controllers
         [Route("Update/{id}")]
         public async Task <IActionResult> Update(UpdateQuestionDto question)
         {
+            var token = Request.Cookies["token"];
+            if (string.IsNullOrEmpty(token))
+            {
+                return RedirectToAction("Login");
+            }
+
             var client = _httpClientFactory.CreateClient();
+            client.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue("Bearer", token);
+
             var payload = new
             {
                 request = question
@@ -92,7 +129,15 @@ namespace IyiOlusAdminPanel.Controllers
         [Route("Delete/{id}")]
         public async Task<IActionResult> Delete(Guid id)
         {
+            var token = Request.Cookies["token"];
+            if (string.IsNullOrEmpty(token))
+            {
+                return RedirectToAction("Login");
+            }
+
             var client = _httpClientFactory.CreateClient();
+            client.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue("Bearer", token);
+
             var response = await client.DeleteAsync(Constants.ApiRoot + $"Questions/{id}");
             return RedirectToAction("Index", "Question");
         }
